[WIP] build: add -macos_version_min to macOS hardened LDFLAGS
This might actually be better placed in depends, as older versions of ld
expect a slightly different version of this flag. macosx_version_min.

man ld:
-macos_version_min version
This is set to indicate the oldest macOS version that that the output is to be
used on.  Specifying a later version enables the linker to assumes features of
that OS in the output file.  The format of version is a macOS version number
such as 10.9 or 10.14