build: Use Clang toolset rather than Darwin for depends Boost
As of Boost 1.72.0, the darwin toolset has been removed from Boost
Build, in favour of the Clang toolset; see this commit:
https://github.com/boostorg/build/commit/eb8da52df3a43ca9616a7a6e11e0ad6fe896c031.

The reasoning (not given in the commit) being that darwin was basically
a copy of the clang toolset anyways. Some comments from the commit
author here: https://github.com/boostorg/build/issues/528.

This change may also resolve the macOS issues seen in #18820.